feat(autoware_tensorrt_vad): move logger to ros_vad_logger.hpp
